+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 
 ### BC breaks
 
+  * Removed `stubbles\ioc\App::createModeBindingModule($projectPath, $mode = null)`
+`   * project path and mode are now bound automatically if not explicitly specified
+    * to overrule or configure the defaults use `stubbles\ioc\App::runtime($mode = null)` instead
+  * Deprecated `stubbles\ioc\App::bindCurrentWorkingDirectory()`, use `stubbles\ioc\App::currentWorkingDirectory()` instead, will be removed with 6.0.0
+  * Deprecated `stubbles\ioc\App::bindHostname()`, use `stubbles\ioc\App::hostname()` instead, will be removed with 6.0.0
   * Removed possibility to change values on annotations, annotations should be read only.
   * It is now possible to have more than one annotation of the same type. Retrieving only one annotation via one of the following methods will only return the first defined one:
     * `stubbles\lang\reflect\ReflectionClass::annotation()`